requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
bleach==6.0.0
cffi==1.15.1
cryptography==40.0.2
Deprecated==1.2.13
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
django-recaptcha==3.0.0
djangorestframework==3.14.0
importlib-metadata==6.6.0
jwcrypto==1.4.2
Markdown==3.4.3
pkg_resources==0.0.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
uWSGI==2.0.21
webencodings==0.5.1
wrapt==1.15.0
zipp==3.15.0
